Clarify the requisites to use `:tx-data-fn` migrations
The example didn't show the namespace in which the `:tx-data-fn` function lived, and it wasn't clear from the description that the value of the `:tx-data-fn` key is a fully qualified function name. So make it clear, and updated the example to show it in plain sight. Also make sure we clearly state that the namespace must be available in the application classpath, as we try to `require` it.

Finally fix the path of the migration file in the example, as it didn't match the namespace of the function to be invoked.