Throw exception if migration cannot be transacted
We were throwing ExceptionInfo exceptions in case the Datomic
functions failed, but not when the user migration functions were
failing. In fact, we were ignoring those errors and not reporting them
back at all.

So make sure we throw exceptions in all cases, including the
underlying details. And document that the main entry point to the
migrations machinery can throw those exceptions, and that the user
should check for them.